+++ src/index.js
@@ -17,13 +17,14 @@ import secrets from '@adobe/helix-shared-secrets';
 import apex from './apex/handler.js';
 import cwv from './cwv/handler.js';
 import notFoundHandler from './notfound/handler.js';
-import notFoundDigestHandler from './notfoundigest/handler.js';
+import notFoundDigestHandler from './notfoundigest/handler-orgs.js';
 
 export const HANDLERS = {
   apex,
   cwv,
   404: notFoundHandler,
-  '404-digest': notFoundDigestHandler,
+  '404-digest-orgs': notFoundDigestHandler,
+  '404-digest-sites': notFoundDigestHandler,
 };
 
 function guardEnvironmentVariables(fn) {
